大体整理了一下ino, size ,offset 的类型为64位.

+++ trunk/src/client/osd_conn.h
@@ -1,9 +1,12 @@
 
-struct write_buf* write_buf_hash_find(uint32_t ino) ;
+#include "protocol.gen.h"
+#include "protocol.h"
+
+struct write_buf* write_buf_hash_find(uint64_t ino) ;
 
 
 
-void buffered_write(struct file_stat * stat, uint64_t offset, uint32_t size, const uint8_t * buff);
+void buffered_write(struct file_stat * stat, uint64_t offset, uint64_t size, const uint8_t * buff);
 
 int buffered_write_flush(struct file_stat * stat);
